requirements.txt
bcrypt==3.2.0
Flask==1.1.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.23
PyJWT==2.0.1
gunicorn